pyproject.toml
[tool.poetry]
name = "turret_python_interface"
version = "0.1.0"
description = ""
authors = ["Joshua Salzedo <joshuasalzedo@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<4"
attrs = "^21.2.0"
cattrs = "^1.7.1"
cobs = "^1.1.4"
crc = "^1.0.1"
pyserial = "^3.5"
loguru = "^0.5.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
ipython = "^7.26.0"
black = "^21.7b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
